data/problems/9af7be61-46f0-45b6-a40f-88e456663e11/extracted/templar-full/templar-full/requirements.txt
pytest>=7.0
python-dateutil>=2.0
